+++ packages/webpack-plugin/src/index.ts
@@ -1,6 +1,6 @@
 export * from './plugin';
 export * from './plugin-utils';
 export * from './loader-utils';
-export * from './override-config';
+export * from './webpack-config-stylable-excludes';
 export * from './types';
 export * from './webpack-loader-types';
